+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.4.3
 
+- Support tracking Throwable and Error, not only Exceptions.
 - Support for internal context and override of SDK version.
 
 ## 0.4.2